Extract common scripting functions
