+++ spdivik/__main__.py
@@ -1,61 +1,16 @@
 #!/usr/bin/env python
-import argparse as _agp
 import json
 import logging
 import multiprocessing
 import os
 import pickle
-import time
 import typing
 import numpy as np
-import scipy.io as scio
 from tqdm import tqdm
 import spdivik.predefined as pred
 import spdivik.summary as _smr
 import spdivik.types as ty
-from spdivik import __version__
-
-
-def parse_args():
-    parser = _agp.ArgumentParser()
-    parser.add_argument('--source', '-s',
-                        help='Source file with data. Should contain'
-                             ' observations in rows and features in columns.',
-                        action='store', dest='source', required=True)
-    parser.add_argument('--destination', '-d',
-                        help='Destination directory into which results will be'
-                             ' saved.',
-                        action='store', dest='destination', required=True)
-    parser.add_argument('--config', '-c',
-                        help='Configuration file for the experiment.',
-                        action='store', dest='config', required=True)
-    return parser.parse_args()
-
-
-def prepare_destination(destination: str) -> str:
-    datetime = time.strftime("%Y%m%d-%H%M%S")
-    destination = os.path.join(destination, datetime)
-    os.makedirs(destination)
-    return destination
-
-
-def setup_logger(destination: str):
-    log_destination = os.path.join(destination, 'logs.txt')
-    log_format = '%(asctime)s [%(levelname)s] %(filename)40s:%(lineno)3s' \
-                 + ' - %(funcName)40s\t%(message)s'
-    logging.basicConfig(format=log_format, level=logging.DEBUG, filemode='w',
-                        filename=log_destination)
-    logging.info("Using spdivik, version " + __version__)
-
-
-def load_config(config_path, destination: str):
-    logging.info("Reading configuration file: " + config_path)
-    with open(config_path) as configfile:
-        config = json.load(configfile)
-    logging.info('Backing up configuration.')
-    with open(os.path.join(destination, 'config.json'), 'w') as configfile:
-        json.dump(config, configfile)
-    return config
+import spdivik._scripting as sc
 
 
 def build_experiment(config, data: np.ndarray) -> typing.Tuple[pred.Divik, tqdm]:
@@ -73,29 +28,6 @@ def build_experiment(config, data: np.ndarray) -> typing.Tuple[pred.Divik, tqdm]
     return pred.scenarios[scenario](pool=pool,
                                     progress_reporter=progress_reporter,
                                     **config), progress_reporter
-
-
-def _load_mat(path: str) -> np.ndarray:
-    data = scio.loadmat(path)
-    assert len(data) == 1, \
-        'There should be a single variable inside MAT-file: ' + path
-    return np.array(list(data.values())[0])
-
-
-def load_data(path: str) -> np.ndarray:
-    logging.info("Loading data: " + path)
-    path = path.lower()
-    if path.endswith('.csv') or path.endswith('.txt'):
-        loader = np.loadtxt
-    elif path.endswith('.npy'):
-        loader = np.load
-    elif path.endswith('.mat'):
-        loader = _load_mat
-    else:
-        message = 'Unsupported data format: ' + os.path.splitext(path)[1]
-        logging.error(message)
-        raise IOError(message)
-    return loader(path)
 
 
 def _make_summary(result: typing.Optional[ty.DivikResult]):
@@ -137,11 +69,11 @@ def save(result: typing.Optional[ty.DivikResult], destination: str):
 
 
 def main():
-    arguments = parse_args()
-    destination = prepare_destination(arguments.destination)
-    setup_logger(destination)
-    config = load_config(arguments.config, destination)
-    data = load_data(arguments.source)
+    arguments = sc.parse_args()
+    destination = sc.prepare_destination(arguments.destination)
+    sc.setup_logger(destination)
+    config = sc.load_config(arguments.config, destination)
+    data = sc.load_data(arguments.source)
     experiment, progress = build_experiment(config, data)
     progress.total = data.shape[0]
     progress.update(0)
